YAY-21 fix: resolve merge conflict
